Process subscriptions using queues instead of direct access
Apsis may cause timeouts under heavy load. This allows us to retry
subscriptions even if they fail the first time.

This is enabled by injection of an Apsis client and a queue which new
subscriptions are added to. A queue worker will then take subscriptions
from the queue and pass them to Apsis.

We also change the processing order. Instead of adding messages from
the response to the form state we use a set of fixed strings and use
these only for Drupal messages. Clients can then call drupal_get_message
to retrieve these.