Support injection of Apsis client HTTP client and config
This will allow for using a HTTP client with custom configuration when
preparing for queuing new subscribers.

This should also make the client more testable down the line.

It it not necessary to include the configuration in the injection but
it was added here nonetheless to improve the code even more.

SAMD8-567