Use client with low timeout for requests before queuing subscriptions
This will allow us to deal with slow responses even better.

If the requests complete within time we can provide proper feedback to
The user in regard to whether she is already subscribed to the list
before we queue the actual subscription.
If they time out the API could be under pressure and will respond
slowly. If we cannot get a response fast enough we act as if the user
is a new subscriber when sending feedback before queuing. That may not 
be correct but the most important part is that we queue the potentially
new subscription before timing out. Apsis can handle existing
subscribers anyway.

SAMD8-567