Improve error handling by introducing exceptions.
Currently the Apsis class does not allow the caller to distinguish
Between different types of errors which may occur in the communication
with the APSIS API. All exceptions are caught and FALSE returned.

This change introduces a mapper which handles the rather complex logic
Of mapping from generic HTTP request exceptions to specific APSIS
Exceptions based on custom APSIS error codes and messages as well as
HTTP status codes.

API callers are updated to handle the new exceptions and act
accordingly.

Most importantly this means that we can handle queued subscriptions
correctly. If an email is on an opt-out list or the email is invalid
then there is no need to try to process the subscription at a later
point in time. Thus we avoid blocking the queue.